pyproject.toml
[tool.poetry]
name = "lokidb_sdk"
version = "0.1.0"
description = "LokiDB python client lib"
authors = ["yehoyada <hvuhsg5@gmail.com>"]
readme = "README.md"
packages = [{include = "lokidb_sdk"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
grpcio = "^1.48.1"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.48.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
